user api

+++ src/main/java/com/cghue/projecthousemaidwebapp/domain/OrderDetail.java
@@ -6,8 +6,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.List;
-
 @Entity(name = "order_details")
 @AllArgsConstructor
 @NoArgsConstructor
@@ -30,6 +28,6 @@ public class OrderDetail {
 
     private Integer quantity;
 
-    private Integer houseSize;
+    private Double houseSize;
 
 }
